Spectra/requirements.txt
asyncpraw==7.8.1
discord.py==2.5.0
python-dotenv==1.0.1
Requests==2.32.3
